Move dependencies module indo conda store server module
Only the fastapi app needs to define these depends pieces. Other
pieces of the application should not be accessing this module.
