+++ conda-store-server/conda_store_server/_internal/server/views/api.py
@@ -15,7 +15,7 @@ from conda_store_server import __version__, api, app
 from conda_store_server._internal import orm, schema, utils
 from conda_store_server._internal.environment import filter_environments
 from conda_store_server._internal.schema import AuthenticationToken, Permissions
-from conda_store_server.server import dependencies
+from conda_store_server._internal.server import dependencies
 from conda_store_server.server.auth import Authentication
