+++ conda-store-server/conda_store_server/_internal/server/views/conda_store_ui.py
@@ -4,7 +4,7 @@
 
 from fastapi import APIRouter, Depends, Request
 
-from conda_store_server.server import dependencies
+from conda_store_server._internal.server import dependencies
 
 router_conda_store_ui = APIRouter(tags=["conda-store-ui"])
