+++ conda-store-server/conda_store_server/_internal/server/views/metrics.py
@@ -6,7 +6,7 @@ from fastapi import APIRouter, Depends
 from fastapi.responses import PlainTextResponse
 
 from conda_store_server import api
-from conda_store_server.server import dependencies
+from conda_store_server._internal.server import dependencies
 
 router_metrics = APIRouter(tags=["metrics"])
