+++ conda-store-server/conda_store_server/_internal/server/views/ui.py
@@ -13,7 +13,7 @@ from conda_store_server._internal.action.generate_constructor_installer import (
     get_installer_platform,
 )
 from conda_store_server._internal.schema import Permissions
-from conda_store_server.server import dependencies
+from conda_store_server._internal.server import dependencies
 
 router_ui = APIRouter(tags=["ui"])
