+++ conda-store-server/tests/_internal/server/views/test_api.py
@@ -15,7 +15,7 @@ from fastapi.testclient import TestClient
 
 from conda_store_server import CONDA_STORE_DIR, __version__
 from conda_store_server._internal import schema
-from conda_store_server.server import dependencies
+from conda_store_server._internal.server import dependencies
 
 
 @contextlib.contextmanager
